base class and refactored tests
