+++ agrifoodpy/food/food.py
@@ -124,195 +124,7 @@ def FoodSupply(items, years, quantities, regions=None, elements=None,
     return fbs
 
 @xr.register_dataset_accessor("fbs")
-class FoodBalanceSheet:
-    def __init__(self, xarray_obj):
-        self._validate(xarray_obj)
-        self._obj = xarray_obj
-
-    @staticmethod
-    def _validate(obj):
-        """Validate fbs xarray, checking it is a DataSet and has the minimum set
-        of arrays
-        """
-        if not isinstance(obj, xr.Dataset):
-            raise TypeError("Food Balance Sheet must be an xarray.DataSet")
-
-        # required_data_arrays = ["production", "imports", "exports", "food"]
-        # missing_data_arrays = []
-
-        # for data_array_name in required_data_arrays:
-        #     if data_array_name not in obj.data_vars:
-        #         missing_data_arrays.append(data_array_name)
-
-        # if missing_data_arrays:
-        #     raise AttributeError(f"Missing data arrays: {missing_data_arrays}")
-
-
-    def add_items(self, items, copy_from=None):
-        """Extends the item list of a food balance sheet according to the
-        defined input item list
-
-        Parameters
-        ----------
-        fbs : xarray.Dataset
-            Input food balance sheet Dataset
-        items : list, int, string
-            list of item names to be added to the data
-        copy_from : list, int, string, optional
-            If provided, this is the list of items already on the food balance
-            sheet to copy data from.
-        
-        Returns
-        -------
-        fbs : xarray.Dataset
-            FAOSTAT formatted Food Supply dataset with new items added.
-        """
-
-        fbs = self._obj
-
-        # Check for duplicates
-        indexes = np.unique(items, return_index=True)[1]
-        items = [items[index] for index in sorted(indexes)] #issues with np.unique
-        
-        new_items = xr.DataArray(data = np.ones(len(items)),
-                                coords = {"Item":items})
-
-        if copy_from is not None:
-            fbs_pivot = fbs.sel(Item=copy_from)
-            fbs_pivot["Item"] = items
-            new_fbs = fbs_pivot*new_items
-            
-        else:
-            new_items *= np.nan
-            new_fbs = fbs.isel(Item=0)*new_items
-            
-        # Concatenate to input array
-        concat_fbs = xr.concat([fbs, new_fbs], dim="Item") 
-            
-        return concat_fbs
-
-    def add_years(self, years, projection="empty"):
-        """Extends the year range of a food balance sheet according to the
-        defined maximum year
-
-        Parameters
-        ----------
-        fbs : xarray.Dataset
-            Input food balance sheet Dataset
-        years : list, int
-            list of years to be added to the data
-        projection : string or array_like
-            Projection mode. If "constant", the last year of the input food
-            balance sheet is copied to every new year. If "empty", values are
-            initialized and set to zero. If a float array is given, these are
-            used to populate the new year using a scaling of the last year of
-            the array
-
-        Returns
-        -------
-        fbs : xarray.Dataset
-            FAOSTAT formatted Food Supply dataset with new years added.
-        """
-
-        fbs = self._obj
-
-        indexes = np.unique(years, return_index=True)[1]
-        years = [years[index] for index in sorted(indexes)]
-        
-        if projection == "empty":
-            data = np.zeros(len(years))*np.nan
-        elif projection == "constant":
-            data = np.ones(len(years))
-        else:
-            data = np.ones(len(years)) * projection
-        
-        new_years = xr.DataArray(data=data,
-                                    coords = {"Year":years})
-
-        # Select last year as pivot
-        fbs_pivot = fbs.isel(Year=-1)
-        
-        # Create DS or DA by multiplying along last value
-        new_fbs = fbs_pivot*new_years
-        
-        # Concatenate to input array
-        concat_fbs = xr.concat([fbs, new_fbs], dim="Year") 
-            
-        return concat_fbs
-
-    def add_regions(self, regions, copy_from=None):
-        """Extends the region list of a food balance sheet according to the
-        defined input region list
-
-        Parameters
-        ----------
-        fbs : xarray.Dataset
-            Input food balance sheet Dataset
-        new_regions : list, int, string
-            list of region names to be added to the data
-        copy_from : list, int, string
-            If provided, this is the list of regions already on the food balance
-            sheet to copy data from.
-        labels : dict 
-            Dictionary containing the new label for the regions matched to its
-            corresponding label coordinate.
-        
-        Returns
-        -------
-        fbs : xarray.Dataset
-            FAOSTAT formatted Food Supply dataset with new regions added.
-        """
-        
-        fbs = self._obj
-
-        indexes = np.unique(regions, return_index=True)[1]
-        regions = [regions[index] for index in sorted(indexes)]
-        
-        new_regions = xr.DataArray(data = np.ones(len(regions)),
-                                coords = {"Region":regions})
-
-        if copy_from is not None:
-            fbs_pivot = fbs.sel(Region=copy_from)
-            fbs_pivot["Region"] = regions
-            new_fbs = fbs_pivot*new_regions
-            
-        else:
-            new_regions *= np.nan
-            new_fbs = fbs.isel(Region=0)*new_regions
-            
-        # Concatenate to input array
-        concat_fbs = xr.concat([fbs, new_fbs], dim="Region") 
-            
-        return concat_fbs
-    
-    def group_sum(self, coordinate, new_name=None):
-        """Sums quantities over items of a equal group labels and, optionally,
-        renames the groups label coordinate.
-
-        Parameters
-        ----------
-        fbs : xarray.Dataset
-            Input food balance sheet Dataset
-        coordinate : string
-            Coordinate name to group items and sum over them 
-        new_name : string, optional 
-            New name for the collapsed coordinate
-        
-
-        Returns
-        -------
-        fbs : xarray.Dataset
-            FAOSTAT formatted Food Supply dataset with new coordinate base.
-        """
-        
-        fbs = self._obj
-
-        grouped_fbs = fbs.groupby(coordinate).sum()
-
-        if new_name is not None:
-            grouped_fbs = grouped_fbs.rename({coordinate:new_name})
-
-        return grouped_fbs
+class FoodBalanceSheet(_XarrayAccessorBase):
 
     def scale_element(self, element, scale, items=None):
         """Scales list of items from an element in a food balance sheet like
@@ -615,31 +427,8 @@ class FoodBalanceSheet:
         return ax
 
 @xr.register_dataarray_accessor("fes")
-class FoodElementSheet:
-    def __init__(self, xarray_obj):
-        self._validate(xarray_obj)
-        self._obj = xarray_obj
-
-    @staticmethod
-    def _validate(obj):
-        """Validate fbs xarray, checking it is a DataArray and has the minimum
-        set of arrays
-        """
-        if not isinstance(obj, xr.DataArray):
-            raise TypeError("Food Balance Sheet must be an xarray.DataSet")
+class FoodElementSheet(_XarrayAccessorBase):
         
-    def add_regions(self, regions, copy_from=None):
-        # use same function as fbs?
-        pass
-
-    def add_years(self, years, projection="empty"):
-        # use same function as fbs?
-        pass
-
-    def add_items(self, items, copy_from=None):
-        # use same function as fbs?
-        pass
-
     def plot_years(self, show="Item", ax=None, colors=None, labels=None,
                    stack=True, **kwargs):
         """ Fill plot with quantities at each year value
